requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
python-dotenv
python-jose[cryptography]
passlib[argon2]
python-multipart
pydantic
email-validator
APScheduler
httpx
psycopg2-binary
aiofiles
